fix (TMS-22): fixed Level indicator in the repository, entity and service

+++ spring-app/src/main/java/org/example/tmsserver/controller/LevelController.java
@@ -19,7 +19,7 @@ public class LevelController {
     }
 
     @PostMapping("/{regionId}/calculate")
-    public ResponseEntity<Level> calculate(@PathVariable Long regionId) {
+    public ResponseEntity<Level> calculate(@PathVariable Integer regionId) {
         Level level = levelService.calculateLevelForRegion(regionId);
         return ResponseEntity.ok(level);
     }
